+++ Sources/ProjectDescription/Package.swift
@@ -54,6 +54,7 @@ extension Package {
     /// - Parameters:
     ///     - url: The valid Git URL of the package.
     ///     - version: The minimum version requirement.
+    @available(*, deprecated, message: "Use Dependencies.swift instead")
     public static func package(url: String, from version: Version) -> Package {
         .package(url: url, .upToNextMajor(from: version))
     }
@@ -63,6 +64,7 @@ extension Package {
     /// - Parameters:
     ///     - url: The valid Git URL of the package.
     ///     - requirement: A dependency requirement. See static methods on `Package.Dependency.Requirement` for available options.
+    @available(*, deprecated, message: "Use Dependencies.swift instead")
     public static func package(url: String, _ requirement: Package.Requirement) -> Package {
         .remote(url: url, requirement: requirement)
     }
@@ -78,6 +80,7 @@ extension Package {
     /// - Parameters:
     ///     - url: The valid Git URL of the package.
     ///     - range: The custom version range requirement.
+    @available(*, deprecated, message: "Use Dependencies.swift instead")
     public static func package(url: String, _ range: Range<Version>) -> Package {
         .remote(url: url, requirement: .range(from: range.lowerBound, to: range.upperBound))
     }
@@ -93,6 +96,7 @@ extension Package {
     /// - Parameters:
     ///     - url: The valid Git URL of the package.
     ///     - range: The closed version range requirement.
+    @available(*, deprecated, message: "Use Dependencies.swift instead")
     public static func package(url: String, _ range: ClosedRange<Version>) -> Package {
         // Increase upperbound's patch version by one.
         let upper = range.upperBound
@@ -112,6 +116,7 @@ extension Package {
     /// on multiple tightly coupled packages.
     ///
     /// - Parameter path: The path of the package.
+    @available(*, deprecated, message: "Use Dependencies.swift instead")
     public static func package(path: Path) -> Package {
         .local(path: path)
     }