+++ Sources/TuistGenerator/Generator/SchemeDescriptorsGenerator.swift
@@ -266,7 +266,10 @@ final class SchemeDescriptorsGenerator: SchemeDescriptorsGenerating {
                           rootPath: AbsolutePath,
                           generatedProjects: [AbsolutePath: GeneratedProject]) throws -> XCScheme.TestAction?
     {
-        guard let testAction = scheme.testAction else { return nil }
+        // Use empty action if nil, otherwise Xcode will create it anyway
+        let testAction = scheme.testAction ?? .empty(
+            withConfigurationName: scheme.runAction?.configurationName ?? BuildConfiguration.debug.name
+        )
 
         var testables: [XCScheme.TestableReference] = []
         var preActions: [XCScheme.ExecutionAction] = []
@@ -863,5 +866,25 @@ final class SchemeDescriptorsGenerator: SchemeDescriptorsGenerating {
         default:
             return nil
         }
+    }
+}
+
+extension TestAction {
+    fileprivate static func empty(withConfigurationName configurationName: String) -> Self {
+        .init(
+            targets: [],
+            arguments: nil,
+            configurationName: configurationName,
+            attachDebugger: true,
+            coverage: false,
+            codeCoverageTargets: [],
+            expandVariableFromTarget: nil,
+            preActions: [],
+            postActions: [],
+            diagnosticsOptions: [],
+            language: nil,
+            region: nil,
+            testPlans: nil
+        )
     }
 }