+++ Sources/TuistLoader/Models+ManifestMappers/Headers+ManifestMapper.swift
@@ -11,59 +11,69 @@ extension TuistGraph.Headers {
     /// - Parameters:
     ///   - manifest: Manifest representation of Headers.
     ///   - generatorPaths: Generator paths.
-    static func from(manifest: ProjectDescription.Headers, generatorPaths: GeneratorPaths) throws -> TuistGraph.Headers {
-        func resolveExcluding(_ excluding: [Path]?) throws -> Set<AbsolutePath> {
-            guard let excluding = excluding else { return [] }
-            var result: Set<AbsolutePath> = []
-            try excluding.forEach { path in
-                let resolved = try generatorPaths.resolve(path: path).pathString
-                let absolute = AbsolutePath(resolved)
-                let globs = AbsolutePath(absolute.dirname).glob(absolute.basename)
-                result.formUnion(globs)
-            }
-            return result
-        }
-
-        func unfoldGlob(_ path: AbsolutePath, excluding: Set<AbsolutePath>) -> [AbsolutePath] {
-            FileHandler.shared.glob(AbsolutePath.root, glob: String(path.pathString.dropFirst())).filter {
-                guard let fileExtension = $0.extension else {
-                    return false
-                }
-                return TuistGraph.Headers.extensions.contains(".\(fileExtension)") && !excluding.contains($0)
-            }
+    static func from(
+        manifest: ProjectDescription.Headers,
+        generatorPaths: GeneratorPaths,
+        productName: String?
+    ) throws -> TuistGraph.Headers {
+        let resolvedUmbrellaPath = try manifest.umbrellaHeader.map { try generatorPaths.resolve(path: $0) }
+        let headersFromUmbrella = try resolvedUmbrellaPath.map {
+            Set(try UmbrellaHeaderHeadersExtractor.headers(from: $0, for: productName))
         }
 
         var autoExlcudedPaths = Set<AbsolutePath>()
-        let publicHeaders: [AbsolutePath]
+        var publicHeaders: [AbsolutePath]
         let privateHeaders: [AbsolutePath]
         let projectHeaders: [AbsolutePath]
 
-        func resolveHeaders(_ list: FileList?) throws -> [AbsolutePath] {
+        let allowedExtensions = TuistGraph.Headers.extensions
+        func unfold(_ list: FileList?,
+                    isPublic: Bool = false) throws -> [AbsolutePath]
+        {
             guard let list = list else { return [] }
             return try list.globs.flatMap {
-                unfoldGlob(
-                    try generatorPaths.resolve(path: $0.glob),
-                    excluding: (try resolveExcluding($0.excluding)).union(autoExlcudedPaths)
-                )
+                try $0.unfold(generatorPaths: generatorPaths) { path in
+                    guard let fileExtension = path.extension,
+                          allowedExtensions.contains(".\(fileExtension)"),
+                          !autoExlcudedPaths.contains(path)
+                    else {
+                        return false
+                    }
+                    if isPublic, let headersFromUmbrella = headersFromUmbrella {
+                        return headersFromUmbrella.contains(path.basename)
+                    }
+                    return true
+                }
             }
         }
 
         switch manifest.exclusionRule {
         case .projectExcludesPrivateAndPublic:
-            publicHeaders = try resolveHeaders(manifest.public)
+            publicHeaders = try unfold(manifest.public, isPublic: true)
+            // be sure, that umbrella was not added before
+            if let resolvedUmbrellaPath = resolvedUmbrellaPath,
+               !publicHeaders.contains(resolvedUmbrellaPath)
+            {
+                publicHeaders.append(resolvedUmbrellaPath)
+            }
             autoExlcudedPaths.formUnion(publicHeaders)
-            privateHeaders = try resolveHeaders(manifest.private)
+            privateHeaders = try unfold(manifest.private)
             autoExlcudedPaths.formUnion(privateHeaders)
-            projectHeaders = try resolveHeaders(manifest.project)
+            projectHeaders = try unfold(manifest.project)
 
         case .publicExcludesPrivateAndProject:
-            projectHeaders = try resolveHeaders(manifest.project)
+            projectHeaders = try unfold(manifest.project)
             autoExlcudedPaths.formUnion(projectHeaders)
-            privateHeaders = try resolveHeaders(manifest.private)
+            privateHeaders = try unfold(manifest.private)
             autoExlcudedPaths.formUnion(privateHeaders)
-            publicHeaders = try resolveHeaders(manifest.public)
+            publicHeaders = try unfold(manifest.public, isPublic: true)
+            // be sure, that umbrella was not added before
+            if let resolvedUmbrellaPath = resolvedUmbrellaPath,
+               !publicHeaders.contains(resolvedUmbrellaPath)
+            {
+                publicHeaders.append(resolvedUmbrellaPath)
+            }
         }
-
         return Headers(public: publicHeaders, private: privateHeaders, project: projectHeaders)
     }
 }