+++ Sources/TuistSupport/Logging/DetailedLogHandler.swift
@@ -35,7 +35,8 @@ public struct DetailedLogHandler: LogHandler {
 
         log.append(message.description)
 
-        output(for: level).log(level: level, message: message, metadata: metadata, file: file, function: function, line: line)
+        (output(for: level) as? SuppressedWarningLogHandler)?
+            .log(level: level, message: message, metadata: metadata, file: file, function: function, line: line)
     }
 
     func output(for level: Logger.Level) -> LogHandler {
@@ -73,3 +74,17 @@ func timestamp() -> String {
         }
     }
 }
+
+// Protocol needed to suppress the warning because we don't have a valid `source` to pass to the `log` method
+protocol SuppressedWarningLogHandler {
+    func log(
+        level: Logging.Logger.Level,
+        message: Logging.Logger.Message,
+        metadata: Logging.Logger.Metadata?,
+        file: String,
+        function: String,
+        line: UInt
+    )
+}
+
+extension StreamLogHandler: SuppressedWarningLogHandler {}